+++ R/relocate.R
@@ -40,21 +40,26 @@
 #' df2 <- tibble(a = 1, b = "a", c = 1, d = "a")
 #' df2 %>% relocate(is.numeric, .after = is.character)
 #' df2 %>% relocate(is.numeric, .before = is.character)
-relocate <- function(.data, ..., .before, .after) {
+relocate <- function(.data, ..., .before = NULL, .after = NULL) {
   UseMethod("relocate")
 }
 
 #' @export
-relocate.data.frame <- function(.data, ..., .before, .after) {
+relocate.data.frame <- function(.data, ..., .before = NULL, .after = NULL) {
   to_move <- tidyselect::eval_select(expr(c(...)), .data)
 
-  if (!missing(.before) && !missing(.after)) {
+  .before <- enquo(.before)
+  .after <- enquo(.after)
+  has_before <- !quo_is_null(.before)
+  has_after <- !quo_is_null(.after)
+
+  if (has_before && has_after) {
     abort("Must supply only one of `.before` and `.after`")
-  } else if (!missing(.before) && missing(.after)) {
-    where <- tidyselect::eval_select(enquo(.before), .data)
+  } else if (has_before) {
+    where <- tidyselect::eval_select(.before, .data)
     to_move <- c(setdiff(to_move, where), where)
-  } else if (missing(.before) && !missing(.after)) {
-    where <- tidyselect::eval_select(enquo(.after), .data)
+  } else if (has_after) {
+    where <- tidyselect::eval_select(.after, .data)
     to_move <- c(where, setdiff(to_move, where))
   } else {
     where <- 1L